+++ src/ui/vite.config.ts
@@ -2,8 +2,13 @@ import { defineConfig } from 'vite';
 import vue from '@vitejs/plugin-vue';
 import vuetify from 'vite-plugin-vuetify';
 
+import packageJson from '../../package.json';
+
 // https://vitejs.dev/config/
 export default defineConfig({
+    define: {
+        '__APP_VERSION__': JSON.stringify(packageJson.version)
+    },
     plugins: [
         vue(),
         vuetify({ autoImport: true }),